Работа с ошибками, все поразносил по файлам разным, функцией чек аларм вообще горжусь
